% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allele_freqs.R
\name{allele_freqs}
\alias{allele_freqs}
\title{Compute locus allele frequencies}
\usage{
allele_freqs(X, loci_on_cols = FALSE, m_chunk_max = 1000)
}
\arguments{
\item{X}{The genotype matrix (regular R matrix or BEDMatrix object).
Missing values are ignored in averages.}

\item{loci_on_cols}{If \code{TRUE}, \eqn{X} has loci on columns and individuals on rows; if false (the default), loci are on rows and individuals on columns.
If \eqn{X} is a BEDMatrix object, columns are averaged to yield locus allele frequencies (regardless of the value of \code{loci_on_cols}).}

\item{m_chunk_max}{BEDMatrix-specific, sets the maximum number of loci to process at the time.
If memory usage is excessive, set to a lower value than default (expected only for extremely large numbers of individuals).}
}
\value{
The vector of allele frequencies, one per locus.
Names are set to the locus names, if present.
}
\description{
On a regular matrix, this is essentially a wrapper for colMeans or rowMeans depending on \code{loci_on_cols}.
On a BEDMatrix object, the locus allele frequencies are computed keeping memory usage low.
}
\examples{
# Construct toy data
X <- matrix(
    c(0, 1, 2,
      1, 0, 1,
      1, NA, 2),
    nrow = 3,
    byrow = TRUE
)

# row means
allele_freqs(X)
c(1/2, 1/3, 3/4)

# col means
allele_freqs(X, loci_on_cols = TRUE)
c(1/3, 1/4, 5/6)

}
