% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inv_var_est_bayesian.R
\name{inv_var_est_bayesian}
\alias{inv_var_est_bayesian}
\title{Calculate expectations of inverse variance terms over the posterior distribution of the ancestral allele frequency}
\usage{
inv_var_est_bayesian(p_anc_est, kinship, g)
}
\arguments{
\item{p_anc_est}{A vector of sample allele frequencies (each refered to as \code{x} above)}

\item{kinship}{The mean kinship coefficient of the data}

\item{g}{The exponent of the variance term}
}
\value{
A vector of expectations of \code{1 / ( p * ( 1 - p ) )^g} over the marginalized, unknown true ancestral allelel frequencies \code{p}, weighted by how likely they are in the posterior given each input sample allele frequency.
The length of the vector is the same as the input \code{p_anc_est}, and each element in the output was estimated from each corresponding element in the input only.
}
\description{
This function assumes that each sample allele frequency \code{x} in the input has a Beta distribution with mean \code{p}  and variance p*(1-p)*kinship\verb{, where }p\verb{is the true allele frequency and is unknown, while}kinship\verb{is the mean kinship of the sample and is known. Using this assumption, an expectation over the posterior distribution of}p\verb{is calculated for what we call an inverse variance term, namely}1 / ( p * ( 1 - p ) )^g\verb{, where }p * ( 1 - p )\verb{corresponds to the Bernoulli variance, and}g\verb{is some desired exponent. Note that this function estimates these posteriors each from a single data point}x`.
}
\examples{
# select a relatively high value for example
kinship <- 0.1

# try a grid of values, including edge cases (0,1)
m_loci <- 1000
p_anc_est <- 0 : (m_loci - 1) / (m_loci - 1)

# calculate the desired estimates!
# here g=0.5 estimates `1 / sqrt( p * ( 1 - p ) )`
inv_var_est <- inv_var_est_bayesian( p_anc_est, kinship, 0.5 )

}
\seealso{
\code{\link[=p_anc_est_beta_mle]{p_anc_est_beta_mle()}} for another way to get less biased estimates of the ancestral allele frequencies and related terms.
}
