% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_anc_hat_beta_mle.R
\name{p_anc_est_beta_mle}
\alias{p_anc_est_beta_mle}
\title{Calculate maximum likelihood estimates of allele frequencies from Beta model}
\usage{
p_anc_est_beta_mle(p_anc_est, kinship)
}
\arguments{
\item{p_anc_est}{A vector of sample allele frequencies (each refered to as \code{x} above)}

\item{kinship}{The mean kinship coefficient of the data}
}
\value{
A vector of maximum likelihood estimates of allele frequencies.
The length of the vector is the same as the input \code{p_anc_est}, and each element in the output was estimated from each corresponding element in the input only.
}
\description{
For each sample allele frequency \code{x} in the input, this function calculates the maximum likelihood estimate of the true allele frequency p\verb{assuming that the sample was drawn from a Beta distribution with mean}p\verb{ and variance p*(1-p)*kinship}, where \code{kinship} is the mean kinship of the sample and is known.
Note that, oddly, this function estimates the unknown \code{p} from a single data point \code{x}.
Nevertheless, this procedure results in favorable shrinkage of estimate towards 0.5.
}
\examples{
# select a relatively high value for example
kinship <- 0.1

# try a grid of values, including edge cases (0,1)
m_loci <- 1000
p_anc_est <- 0 : (m_loci - 1) / (m_loci - 1)

# calculate the desired estimates!
p_anc_mle <- p_anc_est_beta_mle( p_anc_est, kinship )

# notice values tend to be shrunk towards 0.5,
# except values near the edges of the range are shrunk less
plot( p_anc_est, p_anc_mle )
abline(0,1)

}
\seealso{
\code{\link[=inv_var_est_bayesian]{inv_var_est_bayesian()}} for another way to get unbiased estimates of a specific class of inverse variance terms of interest in this project.
}
