% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pval_aucpr.R
\name{pval_aucpr}
\alias{pval_aucpr}
\title{Area under the precision-recall curve}
\usage{
pval_aucpr(pvals, causal_indexes)
}
\arguments{
\item{pvals}{The vector of association p-values to analyze.
\code{NA} values are allowed in input, are internally set to 1 (worst score) prior to AUC calculation (to prevent methods to get good AUCs by setting more cases to NA).
Non-\code{NA} values outside of [0,1] will trigger an error.}

\item{causal_indexes}{The vector of causal indexes, defining the true classes used for AUC calculation.
Values of \code{causal_indexes} as returned by \code{sim_trait} work.
There must be at least one causal index and at least one non-causal case.}
}
\value{
The PR AUC scalar value.

However, if the input \code{pvals} is \code{NULL} (taken for case of singular association test, which is rare but may happen), then the returned value is \code{NA}.
}
\description{
Calculates the Precision-Recall (PR) Area Under the Curve (AUC) given a vector of p-values and the true classes (causal (alternative) vs non-causal (null)).
This is a wrapper around the function \code{pr.curve} from the \code{PRROC} package, which actually calculates the AUC (see that for details).
}
\examples{
# simulate truly null p-values, which should be uniform
pvals <- runif(10)
# for toy example, take the first two p-values to be truly causal
causal_indexes <- 1:2
# calculate desired measure
pval_aucpr( pvals, causal_indexes )

}
\seealso{
The function \code{pr.curve} from the \code{PRROC} package, which is used internally by \code{pval_aucpr}.
}
