% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_trait.R
\name{sim_trait}
\alias{sim_trait}
\title{Simulate a complex trait from genotypes}
\usage{
sim_trait(X, m_causal, herit, p_anc, mean_kinship, maf_cut = 0.05)
}
\arguments{
\item{X}{The \eqn{m \times n}{m-by-n} genotype matrix.  This is a numeric matrix consisting of reference allele counts (in \code{c(0,1,2,NA)} for a diploid organism).}

\item{m_causal}{The number of causal loci desired.}

\item{herit}{The desired heritability (proportion of trait variance due to genetics).}

\item{p_anc}{The length-\eqn{m} vector of true ancestral allele frequencies.  Recommended way to scale output parameters to achieve the desired heritability.  Either this or \code{mean_kinship} must be specified.}

\item{mean_kinship}{The mean kinship coefficient across all individuals in the data.  This offers an alternative way to scale output parameters to achieve the desired heritability, as \code{p_anc} is only known for simulated data.  Either this or \code{p_anc} must be specified.}

\item{maf_cut}{The optional minimum allele frequency threshold (default 5\%).  This prevents rare alleles from being causal in the simulation.}
}
\value{
A list containing the simulated trait \code{y} (length \eqn{n}), the vector of causal locus indeces \code{i} (length \eqn{m_causal}), and the locus effect size vector \code{beta} (length \eqn{m}, equals zero at every non-causal locus).
}
\description{
Simulate a complex trait \eqn{y} given a SNP genotype matrix and model parameters (the desired heritability and either the true ancestral allele frequencies used to generate the genotypes or the mean kinship).  Users can choose the number of causal loci and minimum marginal allele frequency requirements for the causal loci.  The code selects random loci to be causal, draws random Normal effect sizes for these loci (scaled appropriately) and random independent non-genetic effects.
Below let there be \eqn{m} loci and \eqn{n} individuals.
}
\details{
In order to scale the locus effect size vector correctly to
}
\examples{
# construct a dummy genotype matrix
X <- matrix(
           data=c(0,1,2,1,2,1,0,0,1),
           nrow=3,
           byrow=TRUE
           )
# made up mean kinship for example
mean_kinship <- 0.1
# create simulated trait and associated data
obj <- sim_trait(X=X, m_causal=1, herit=0.8, mean_kinship=mean_kinship)
# trait vector
obj$y
# randomly-picked causal locus index
obj$i
# locus effect size vector
obj$beta

}
