% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simtrait.R
\docType{package}
\name{simtrait}
\alias{simtrait}
\alias{simtrait-package}
\title{simtrait: a package for simulating complex traits from genotypes}
\description{
This package enables simulation of complex (polygenic and continuous) traits from a simulated or real genotype matrix.
The focus is on controling the mean and covariance structure of the data to yield the desired heritability under arbitrary population structures (any underlying kinship matrix).
The main function is \code{\link{sim_trait}}, which returns the simulated trait and the vector of causal loci (randomly selected) and their effect sizes (randomly drawn and scaled appropriately).
The function \code{\link{cov_trait}} computes the expected covariance matrix of the trait given the model parameters (namely the desired heritability and the true kinship matrix).
}
\details{
The recommended inputs are simulated genotypes with known ancestral allele frequencies.
The \code{bnpsd} package simulates genotypes for admixed individuals, resulting in a complex population structure.

For real data it is necessary to estimate the kinship matrix.
The \code{popkin} package provides high-accuracy kinship estimates.
}
\examples{
# construct a dummy genotype matrix
X <- matrix(
           data = c(0,1,2,1,2,1,0,0,1),
           nrow = 3,
           byrow = TRUE
           )
# made up ancestral allele frequency vector for example
p_anc <- c(0.5, 0.6, 0.2)
# desired heritability
herit <- 0.8
# create simulated trait and associated data
obj <- sim_trait(X = X, m_causal = 2, herit = herit, p_anc = p_anc)
# trait vector
obj$trait
# randomly-picked causal locus indeces
obj$causal_indexes
# locus effect size vector
obj$causal_coeffs

# create a dummy kinship matrix for example
kinship <- matrix(
             data = c(0.6,0.1,0, 0.1,0.6,0.1, 0,0.1,0.6),
             nrow = 3,
             byrow = TRUE
             )
# covariance of simulated traits
V <- cov_trait(kinship = kinship, herit = herit)

}
\author{
\strong{Maintainer}: Alejandro Ochoa \email{alejandro.ochoa@duke.edu}

}
