% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_trait_mvn.R
\name{sim_trait_mvn}
\alias{sim_trait_mvn}
\title{Simulate traits from a kinship matrix under the infinitesimal model}
\usage{
sim_trait_mvn(rep, kinship, herit, mu = 0, sigma_sq = 1, tol = 1e-06)
}
\arguments{
\item{rep}{The number of replicate traits to simulate.
Simulating all you need at once is more efficient than simulating each separately, as the kinship matrix is eigendecomposed each time.}

\item{kinship}{The \eqn{n \times n}{n-by-n} kinship matrix of the individuals to simulate from.}

\item{herit}{The desired heritability (proportion of trait variance due to genetics).}

\item{mu}{The desired parametric mean value of the trait (default zero).
The sample mean of the trait will not be exactly zero, but instead have an expectation of \code{mu} (with potentially large variance depending on the kinship matrix and the heritability).}

\item{sigma_sq}{The desired parametric variance factor of the trait (default 1).
This factor corresponds to the variance of an outbred individual (see \code{\link{cov_trait}}).}

\item{tol}{Tolerance factor for an internal test of positive semi-definiteness of the trait covariance matrix.
Procedure fails if any eigenvalues are smaller than \code{-tol} times the absolute value of the largest eigenvalue.
Increase this value only if you are getting errors but you're sure your covariance matrix (the output of \verb{\link{cov_trait}}) is positive semi-definite.}
}
\value{
A list containing the simulated \code{trait} (length \eqn{n}).
}
\description{
Simulate matrix of trait replicates \eqn{Y} given a kinship matrix and model parameters (the desired heritability, total variance scale, and mean).
Although these traits have the covariance structure of genetic traits, and have heritabilities that can be estimated, they do not have causal loci (an association test against any locus should fail).
Below \eqn{n} is the number of individuals.
}
\examples{
# create a dummy kinship matrix
# make sure it is positive definite!
kinship <- matrix(
             data = c(0.6,0.1,0, 0.1,0.5,0, 0,0,0.5),
             nrow = 3,
             byrow = TRUE
             )
# draw simulated traits (matrix)
traits <- sim_trait_mvn( rep = 10, kinship = kinship, herit = 0.8)
traits

}
